+++ packages/composite/src/__package__.ts
@@ -1,3 +1,3 @@
 export const PKG_NAME = "@hpcc-js/composite";
-export const PKG_VERSION = "2.5.28";
-export const BUILD_VERSION = "2.15.10";
+export const PKG_VERSION = "2.5.29";
+export const BUILD_VERSION = "2.15.11";