+++ packages/graph/src/__package__.ts
@@ -1,3 +1,3 @@
 export const PKG_NAME = "@hpcc-js/graph";
 export const PKG_VERSION = "2.17.0";
-export const BUILD_VERSION = "2.15.10";
+export const BUILD_VERSION = "2.15.11";