+++ packages/layout/src/__package__.ts
@@ -1,3 +1,3 @@
 export const PKG_NAME = "@hpcc-js/layout";
-export const PKG_VERSION = "2.16.24";
-export const BUILD_VERSION = "2.15.10";
+export const PKG_VERSION = "2.16.25";
+export const BUILD_VERSION = "2.15.11";