+++ packages/map/src/__package__.ts
@@ -1,3 +1,3 @@
 export const PKG_NAME = "@hpcc-js/map";
-export const PKG_VERSION = "2.14.27";
-export const BUILD_VERSION = "2.15.10";
+export const PKG_VERSION = "2.14.28";
+export const BUILD_VERSION = "2.15.11";