+++ packages/phosphor/src/__package__.ts
@@ -1,3 +1,3 @@
 export const PKG_NAME = "@hpcc-js/phosphor";
-export const PKG_VERSION = "2.14.7";
-export const BUILD_VERSION = "2.15.10";
+export const PKG_VERSION = "2.14.8";
+export const BUILD_VERSION = "2.15.11";