+++ packages/preact-shim/src/__package__.ts
@@ -1,3 +1,3 @@
 export const PKG_NAME = "@hpcc-js/preact-shim";
 export const PKG_VERSION = "2.13.5";
-export const BUILD_VERSION = "2.15.10";
+export const BUILD_VERSION = "2.15.11";