+++ packages/layout/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.16.25](https://github.com/hpcc-systems/Visualization/compare/@hpcc-js/layout@2.16.19...@hpcc-js/layout@2.16.25) (2020-04-21)
+
+
+
+## 2.15.10 (2020-04-21)
+
+
+
+## 2.15.9 (2020-04-17)
+
+
+
+## 2.15.7 (2020-04-11)
+
+**Note:** Version bump only for package @hpcc-js/layout
+
+
+
+
+
 ## [2.16.24](https://github.com/hpcc-systems/Visualization/compare/@hpcc-js/layout@2.16.19...@hpcc-js/layout@2.16.24) (2020-04-21)