+++ packages/map/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.14.28](https://github.com/hpcc-systems/Visualization/compare/@hpcc-js/map@2.14.22...@hpcc-js/map@2.14.28) (2020-04-21)
+
+
+
+## 2.15.10 (2020-04-21)
+
+
+
+## 2.15.9 (2020-04-17)
+
+
+
+## 2.15.7 (2020-04-11)
+
+**Note:** Version bump only for package @hpcc-js/map
+
+
+
+
+
 ## [2.14.27](https://github.com/hpcc-systems/Visualization/compare/@hpcc-js/map@2.14.22...@hpcc-js/map@2.14.27) (2020-04-21)